food_delivery/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiogram==3.13.0
sqlalchemy[asyncio]==2.0.36
alembic==1.13.0
asyncpg==0.29.0
redis[hiredis]==5.2.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.12
pytest==8.3.0
pytest-asyncio==0.24.0
fakeredis==2.26.0
pytest-cov==5.0.0
